Refactor: update import statements to use PongMatch and kFactor from the new file
