+++ src/server/tunnel/joinMatchmaking.ts
@@ -1,10 +1,9 @@
 import {Client, ClientTunnelMessage} from '#types/Clients';
+import PongMatch, {kFactor} from '#lib/PongMatch';
 import Clients from '#lib/Clients';
 import {FastifyInstance} from 'fastify';
-import PongMatch from '#lib/Match';
 import {RankedClient} from '#types/fastify';
 import SQL from 'sql-template-strings';
-import {kFactor} from '#lib/Match';
 import {randomInt} from 'node:crypto';
 
 export default async function joinMatchmaking(
